+++ db/db.json
@@ -2,91 +2,18 @@
   "users": [
     {
       "id": 1,
-      "name": "Alice Johnson",
-      "email": "alice@example.com",
-      "bio": "Software engineer with a passion for API development",
-      "location": "Tokyo",
-      "website": "https://alice.dev",
+      "name": "Test User",
+      "email": "test@example.com",
+      "bio": "Test user for authentication",
+      "location": "Test Location",
+      "website": "https://test.example.com",
       "roles": [
         "user"
       ],
       "createdAt": "2024-01-15T10:00:00Z",
       "updatedAt": "2024-01-15T10:00:00Z",
-      "passwordHash": "$2b$12$LQv3c1yqBWVHxkd0LHAkCOYz6TtxMQJqhN8/LewdBPj/PlYQx.QaK"
-    },
-    {
-      "id": 2,
-      "name": "Bob Smith",
-      "email": "bob@example.com",
-      "bio": "Full-stack developer and tech enthusiast",
-      "location": "Osaka",
-      "website": "https://bobsmith.blog",
-      "roles": [
-        "user",
-        "readonly-admin"
-      ],
-      "createdAt": "2024-01-16T14:30:00Z",
-      "updatedAt": "2024-01-16T14:30:00Z",
-      "passwordHash": "$2b$12$LQv3c1yqBWVHxkd0LHAkCOYz6TtxMQJqhN8/LewdBPj/PlYQx.QaK"
-    },
-    {
-      "id": 3,
-      "name": "Charlie Brown",
-      "email": "charlie@example.com",
-      "bio": "DevOps engineer and cloud architect",
-      "location": "Kyoto",
-      "website": "https://charlie.cloud",
-      "roles": [
-        "user",
-        "admin"
-      ],
-      "createdAt": "2024-01-17T09:15:00Z",
-      "updatedAt": "2024-01-17T09:15:00Z",
-      "passwordHash": "$2b$12$LQv3c1yqBWVHxkd0LHAkCOYz6TtxMQJqhN8/LewdBPj/PlYQx.QaK"
+      "passwordHash": "$2b$12$jdekxBwxYBDcWrTbAW/qu.gAgXyar5ChKNrGwdXgtYqACYqsjI9Ni"
     }
   ],
-  "microposts": [
-    {
-      "id": 1,
-      "userId": 1,
-      "content": "First micropost for user 1",
-      "createdAt": "2024-01-15T11:00:00Z"
-    },
-    {
-      "id": 2,
-      "userId": 1,
-      "content": "Second micropost for user 1",
-      "createdAt": "2024-01-15T12:00:00Z"
-    },
-    {
-      "id": 3,
-      "userId": 2,
-      "content": "First micropost for user 2",
-      "createdAt": "2024-01-16T15:00:00Z"
-    },
-    {
-      "id": 4,
-      "userId": 3,
-      "content": "Another micropost with great content",
-      "createdAt": "2024-01-17T10:00:00Z"
-    },
-    {
-      "id": 5,
-      "userId": 3,
-      "content": "End-to-end test micropost according to OpenAPI specification",
-      "createdAt": "2025-07-06T10:13:36.088Z"
-    },
-    {
-      "id": 6,
-      "userId": 1,
-      "content": "This is a test micropost",
-      "createdAt": "2025-07-06T10:13:36.854Z"
-    },
-    {
-      "id": 7,
-      "userId": 1,
-      "content": "Test micropost content for OpenAPI validation",
-      "createdAt": "2025-07-06T10:13:40.009Z"
-    }
-  ]
+  "microposts": []
 }
