+++ src/middlewares/index.js
@@ -6,6 +6,7 @@ import errorHandler from './errorHandler.js';
 import requestLogger from './requestLogger.js';
 import { apiLimiter, speedLimiter } from './rateLimiting.js';
 import { validateRequest } from './validation.js';
+import { swaggerMiddleware, swaggerServe } from './swagger.js';
 
 export {
   corsMiddleware,
@@ -16,5 +17,7 @@ export {
   errorHandler,
   apiLimiter,
   speedLimiter,
-  validateRequest
+  validateRequest,
+  swaggerMiddleware,
+  swaggerServe
 };
